pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memoryintelligence"
version = "2.0.1"
description = "Official Python SDK for Memory Intelligence — Verifiable meaning infrastructure"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Memory Intelligence Team", email = "sdk@memoryintelligence.io" }]
requires-python = ">=3.10"
keywords = ["ai", "llm", "memory", "rag", "embeddings", "privacy", "compliance",
            "hipaa", "gdpr", "provenance", "umo"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "cryptography>=42.0.0",
    "python-ulid>=2.0.0",
]

[project.optional-dependencies]
dotenv = [
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.3",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://memoryintelligence.io"
Documentation = "https://docs.memoryintelligence.io"
Repository = "https://github.com/somewhere11/memoryintelligence-python-sdk"
Changelog = "https://github.com/somewhere11/memoryintelligence-python-sdk/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["memoryintelligence*"]
exclude = ["tests*", "examples*"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["D100", "D104"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
